pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["AlexPrunici <prunici.alex@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
djangorestframework = "^3.15.0"
psycopg2-binary = "^2.9.9"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
